Moved OpenCL IntrinsicMath implementations.
