+++ Src/ILGPU.Algorithms/CL/CLContext.cs
@@ -1,6 +1,6 @@
 ﻿// ---------------------------------------------------------------------------------------
 //                                   ILGPU Algorithms
-//                        Copyright (c) 2019-2023 ILGPU Project
+//                        Copyright (c) 2019-2024 ILGPU Project
 //                                    www.ilgpu.net
 //
 // File: CLContext.cs
@@ -28,32 +28,6 @@ namespace ILGPU.Algorithms.CL
     static partial class CLContext
     {
         /// <summary>
-        /// The <see cref="CLMath"/> type.
-        /// </summary>
-        private static readonly Type CLMathType = typeof(CLMath);
-
-        /// <summary>
-        /// Represents the <see cref="CLMath.GenerateMathIntrinsic(CLBackend,
-        /// CLCodeGenerator, IR.Value)"/>
-        /// methods.
-        /// </summary>
-        private static readonly MethodInfo MathCodeGenerator =
-            CLMathType.GetMethod(
-                nameof(CLMath.GenerateMathIntrinsic),
-                AlgorithmContext.IntrinsicBindingFlags)
-            .ThrowIfNull();
-
-        /// <summary>
-        /// Represents the intrinsic representation of the
-        /// <see cref="MathCodeGenerator"/>.
-        /// </summary>
-        private static readonly CLIntrinsic MathCodeGeneratorIntrinsic =
-            new CLIntrinsic(
-                MathCodeGenerator,
-                IntrinsicImplementationMode.GenerateCode)
-            .ThrowIfNull();
-
-        /// <summary>
         /// The <see cref="CLGroupExtensions"/> type.
         /// </summary>
         internal static readonly Type CLGroupExtensionsType = typeof(CLGroupExtensions);
@@ -62,24 +36,6 @@ namespace ILGPU.Algorithms.CL
         /// The <see cref="CLWarpExtensions"/> type.
         /// </summary>
         internal static readonly Type CLWarpExtensionsType = typeof(CLWarpExtensions);
-
-        /// <summary>
-        /// Resolves a CL intrinsic for the given math-function configuration.
-        /// </summary>
-        /// <param name="name">The intrinsic name.</param>
-        /// <param name="types">The parameter types.</param>
-        /// <returns>The resolved intrinsic representation.</returns>
-        private static CLIntrinsic GetMathIntrinsic(string name, params Type[] types)
-        {
-            var targetMethod = CLMathType.GetMethod(
-                name,
-                AlgorithmContext.IntrinsicBindingFlags,
-                null,
-                types,
-                null)
-                .ThrowIfNull();
-            return new CLIntrinsic(targetMethod, IntrinsicImplementationMode.Redirect);
-        }
 
         /// <summary>
         /// Registers an intrinsic mapping.
